Add local client play, config option and bug fix
Fixed the restart bug so that now if the templated tahoe.cfg
changes tahoe will be restarted with or without torsocks according
to the role variable use_torsocks.

Added documentation to explain how to install the tahoe-lafs client locally.

Added the "preferred peers" configuration option to the templated tahoe.cfg
- this is used in Leif's truckee and truckee68 branches to specify which
storage nodes to try to connect to first.